Yielding the same array each time saves a lot of time on garbage collection.
Starting from an array of `remap` elements is reasonably fast compared to reallocating
it in each call to `visit`, which adds about 69% to the run time when the callback/body is trivial.

Starting from an empty array adds about 17% to the runtime (compared to `remap` elements),
probably because the JIT has to emit more array operations.

Prefilling with a constant, though incorrect, is even faster than remap;
maybe the runtime can pick a simpler array subclass? e.g.
- p.map(() => 'why???') as unknown as T[] : runtime is about 95% compared to remap
- p.slice() as unknown as T[]             : runtime is about 93% compared to remap -- fastest